+++ src/prelude.rs
@@ -1,6 +1,6 @@
 //! Contains the most useful imports for writing tests and value generators.
 
 pub use crate::{
-    Codie, Dicetest, Die, DieOnce, Fate, codice, dice, hint, hint_debug, hints, stat, stat_debug,
-    stats,
+    Codie, Dice, Dicetest, Die, DieOnce, Fate, codice, dice, die, hint, hint_debug, hints, stat,
+    stat_debug, stats,
 };